correção enum
git-svn-id: https://desktop-k8125me/svn/B1/B1Base@85 eafdeada-f308-9343-9a48-94a505fd8ed7

+++ B1Base/View/BaseView.cs
@@ -130,7 +130,7 @@ namespace B1Base.View
 
             foreach (var enumValue in enumValues)
             {
-                combo.ValidValues.Add(enumValue.ToString(), Model.EnumOperation.GetEnumDescription(enumValue));
+                combo.ValidValues.Add(((int)enumValue).ToString(), Model.EnumOperation.GetEnumDescription(enumValue));
             }
         }
